requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
python-dotenv>=0.19.0
cryptography>=3.4.0
# 云部署优化
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
aiomysql>=0.1.1
# 监控和日志
prometheus-client>=0.14.0
python-json-logger>=2.0.0
# 健康检查
py-healthcheck>=1.0.0